feat: move dayjs and i18n setup to separate file
